2023/abandoned/tmp24takeout/tmp24.ru/requirements.txt
MySQL-python
Flask
Flask-Assets
Flask-Script
Flask-SQLAlchemy
Flask-Uploads
Flask-WTF
alembic
